backend/requirements.txt
fastapi[standard]
httpx
uvicorn
google.generativeai
# python-doten